I've made the messages and prompts more user friendly and I've adjusted the program to use unsigned integers of 64 bytes to boost the number of possible seeds for encryption.

+++ Cyphox/Cypher.h
@@ -12,7 +12,7 @@ public:
 	Cypher();
 	~Cypher();
 
-	virtual string Encrypt( int seed, string decMsg ) = 0;
-	virtual string Decrypt( int seed, string encMsg ) = 0;
+	virtual string Encrypt( uint64_t seed, string decMsg ) = 0;
+	virtual string Decrypt( uint64_t seed, string encMsg ) = 0;
 };
